requirements.txt
APScheduler
bottle
pycryptodome
python-twitter
numpy
datetime